[FEAT] Worker: Implement get_priority()
The method returns an integer and takes one argument, the time slot you
want to look the priority.

The priority start at 0 and for each hours between the begining and the
time slot it will be increase or decrease.
It is increase by 1 for each time slot where the worker had a shift.
It is decrease by 1 for each time slot where the worker had a break.